Port SPT to ppc64le
This patch ports SPT to ppc64le but does not provide full support for
the thread-local storage support yet, so that test_nospt and test_spt
do not work at this stage.

The ppc64le implementation will need the tls base address, so we get it from
the tender's boot info parameters. We add a parameter uint64_t *tls_base
to platform_init so we can get the base address, which we then pass to
crt_init_tls as parameter.

ppc64 can only protect memory in increments of 64kb size, so the linker
script has to be adjusted accordingly.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>